fix: correct import paths after directory restructuring
Fix import paths that were incorrectly updated in the previous commit.
Ensure all paths correctly reference the new locations in app directory.
