+++ app/activities/components/activity-list.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import Link from "next/link";
-import { UsePagination } from "../components/rotion-wrappers";
-import Hed from "../components/hed";
+import { UsePagination } from "../../components/rotion-wrappers";
+import Hed from "../../components/hed";
 import type { Activity } from "../../lib/activity";
 
 type Props = {
