+++ app/styles/globals.css
@@ -992,7 +992,7 @@ a:focus {
 /* ACCENT */
 .accent {
 	background-color: transparent;
-	background-image: url("../public/bg.svg");
+	background-image: url("../../public/bg.svg");
 	background-position: top -900px right -800px;
 	background-repeat: no-repeat;
 	position: absolute;
